pyproject.toml
[project]
name = "django-enum"
version = "2.3.0"
description = "Full and natural support for enumerations as Django model fields."
requires-python = ">=3.10,<4.0"
authors = [
  {name = "Brian Kohan", email = "bckohan@gmail.com"},
]
license = "MIT"
license-files = [ "LICENSE" ]
repository = "https://github.com/django-commons/django-enum"
homepage = "https://django-enum.rtfd.io"
readme = "README.md"
keywords = [
    "enum",  "properties", "defines", "field", "django", "database",
    "bitmask", "mask", "bitfield", "flags"
]
dependencies = [
    "django>=4.2,<6.1",
]
classifiers = [
    "Environment :: Console",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true


[tool.hatch.build.targets.sdist]
# todo - these artifacts should be moved into tests
exclude = ["benchmarks.json", "plot_benchmarks.py"]

[tool.hatch.build.targets.wheel]
packages = ["src/django_enum"]


[project.optional-dependencies]
properties = [
    "enum-properties>=2.5.1"
]
filters = ["django-filter>=21"]
rest = ["djangorestframework>=3.14,<4.0"]


[project.urls]
"Homepage" = "https://django-enum.readthedocs.io"
"Documentation" = "https://django-enum.readthedocs.io"
"Repository" = "https://github.com/django-commons/django-enum"
"Issues" = "https://github.com/django-commons/django-enum/issues"
"Changelog" = "https://django-enum.readthedocs.io/en/latest/changelog.html"
"Code_of_Conduct" = "https://github.com/django-commons/django-enum/blob/main/CODE_OF_CONDUCT.md"


[dependency-groups]
docs = [
    "docutils>=0.21.2",
    "doc8>=1.1.2",
    "furo>=2024.8.6",
    "readme-renderer[md]>=44.0",
    "sphinx>=8.0.0",
    "sphinx-autobuild>=2021.3.14",
    "sphinx-tabs>=3.4.7",
    "sphinxcontrib-django>=2.5",
    "python-dateutil>=2.9.0.post0",
]
psycopg2 = [
    "psycopg2>=2.9.10",
]
psycopg3 = [
    "psycopg",
]
mysql = [
    "mysqlclient>=1.4.0",
]
cx_oracle = [
    "cx-oracle>=8.3.0",
]
oracledb = [
    "oracledb>=2.3.0",
]

typing = [
    "pyright>=1.1.390",
    "mypy>=1.19.1",
    "django-stubs[compatible-mypy]>=5.1.3",
    "django-stubs-ext>=5.1.3",
    "types-django-filter>=25.2.0.20251228",
    "djangorestframework-stubs>=3.16.1",
    "types-pyyaml>=6.0.12.20241230",
]
lint = [
    "ruff>=0.15.0", 
    "readme-renderer[md]>=43.0",
    "doc8>=1.1.2"
]
precommit = [
    "pre-commit>=3.5.0",
]
coverage = ["coverage>=7.6.1"]
test = [
    { include-group = "coverage" },
    "beautifulsoup4>=4.13.3",
    "django-test-migrations>=1.5.0",
    "playwright>=1.58.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "pytest-env>=1.1.5",
    "pytest-playwright>=0.7.0",
    "python-dateutil>=2.9.0.post0",
]
benchmarks = [
    "matplotlib>=3.9.4",
    "numpy>=2.2; python_version >= '3.13'",
    "numpy>=2.0,<2.2; python_version < '3.13'",
    "tqdm>=4.67.1"
]
dev = [
    { include-group = "typing" },
    { include-group = "test" },
    { include-group = "docs" },
    { include-group = "lint" },
    { include-group = "precommit" },
    "packaging>=24.2",
    "tomlkit>=0.13.3",
    "ipdb>=0.13.13",
    "ipython>=8.18.1"
]

[tool.mypy]
# The mypy configurations: http://bit.ly/2zEl9WI
# allow_redefinition = false
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
mypy_path = ["src"]
files = [
  "src/django_enum/**/*.py",
  "tests/examples/**/*.py"
]

[tool.django-stubs]
django_settings_module = "tests.settings"


[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src/django_enum", "tests/examples/**"]
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = false
reportUnusedFunction = false
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalMemberAccess = false
reportGeneralTypeIssues = false
reportOptionalSubscript = false
reportPrivateImportUsage = false


# todo doc8 configuration here is not being picked up and doesnt seem to be working
# as expected - for now pass these parameters manually
[tool.doc8]
max-line-length = 100
sphinx = true


[tool.pytest.ini_options]
# py.test options:
DJANGO_SETTINGS_MODULE = "tests.settings"
env = [
    "IS_PYTEST_RUN=1"
]
python_files = "test*.py"
norecursedirs = "*.egg .eggs dist build docs .tox .git __pycache__"

addopts = [
  "--strict-markers",
  "--cov=django_enum",
  "--cov-branch",
  "--cov-report=term-missing:skip-covered"
]
markers = [
  "ui: browser-based tests",
  "screenshots: tests that generate screenshots of the admin interface for documentation purposes"
]

[tool.coverage.run]
omit = ["tests/**/*py"]
branch = true
source = ["django_enum"]
concurrency = ["thread", "multiprocessing"]
parallel = true
relative_files = true
command_line = "-m pytest --cov=django_enum"

[tool.coverage.paths]
source = ["django_enum"]


[tool.ruff]
line-length = 88
exclude = [
    "doc",
    "dist",
    "examples",
    "plot_benchmarks.py"
]
target-version = "py310"

[tool.ruff.lint]
exclude = ["tests/**/*"]
